Makefile not needed for testing anymore
