+++ .travis.yml
@@ -5,8 +5,6 @@ node_js:
 
 before_script:
   - npm install --quiet -g grunt-cli karma bower
-  - make &
-  - sleep 5
 
 after_script: killall node
